Move the EL2 mapping away from pmap
We don't need the full pmap to create the mappings for the hypervisor
as they will never fault in. Move the EL2 mappings to be handled by
vmm_mmu.c.

While here add initial support for invalidating the mappings on
cleanup. I'm not sure if this is completely correct, but seems to work
on the N1SDP.