hwt: fix double free/use after free
The error path for hwt_backend_init() results in a double free or use
after free and a kernel panic.

hwt_ctx_free() already calls hwt_free_thread(thr) when the refcount is
zero as well as removing it from the ctx->threads TAILQ, so calling
hwt_ctx_free() alone is correct.