hwt: spe: enable SPE at EL2
Where VHE is implmented and enabled, the kernel is booted at EL2. To
enable SPE at EL2, MDCR_E2PB should be 0b00 and therefore does not
require setting explicitly.

For the non-VHE enabled path, the kernel is booted at EL1 so enable SPE
at EL1 with 0b11.

VHE is mandatory as of Armv8.1 and SPE was introduced in Armv8.2. As
such enabling SPE in the non-VHE implemented path does not make sense.